refactor(frontend)(#1422): split DataPreview.tsx + BottomPanel.tsx into <=500 LOC modules
Wave 2 W2-B (umbrella #1428): behavior-preserving split of the two largest
god files on the bottom-panel surface.

- DataPreview.tsx (1110 -> 126 LOC) split into DataPreview.parts/:
  refEntries.ts, luts.ts, ImageViewer.tsx (further split into
  useImagePanZoom/useLUTProcessing/ImageCanvas/SliceSliderRow/ZoomRow/
  LUTSelector/MinMaxRange), TableViewer.tsx (useTablePageFetch +
  TableHeader/TableBody/PaginationControls), PreviewRenderer.tsx (per-kind
  render helpers), useSlicePreview.ts (#899 hook), useOmeMetadata.ts
  (ADR-043 FR-013 hook).

- BottomPanel.tsx (1030 -> 108 LOC) split into BottomPanel.parts/:
  CaretPreservingTextInput.tsx (#710), FormatCapabilityConfig.tsx,
  codeBlockPorts.ts (pure helpers), CodeBlockConfigEditor.tsx (form +
  CodeBlockEnvironmentEditor + CodeBlockPortTable), ConfigPanel.tsx
  (with EnumField/ScalarField/ConfigField + runBrowseDialog),
  LogViewer.tsx, TabBar.tsx.

Every produced file is <=500 LOC. Public exports (DataPreview,
extractRefEntries, RefEntry, BottomPanel) and every data-testid marker
preserved. Three eqeqeq violations previously tolerated by GOD_FILE waivers
are rewritten to strict equality. The eslint.config.js waiver list is
untouched on this PR; Wave 4 strips the entries after all three Wave 2
PRs land.

Tests: 504 pass / 13 skipped (added 23 new cases:
BottomPanel.parts/codeBlockPorts.test.ts +15, DataPreview.parts/
refEntries.test.ts +8). npm run lint (0 errors, 46 warnings, down from
47), format:check, typecheck, build all clean. Sentrux MCP free-tier:
3/15 rules pass, quality_signal=4442 unchanged. Chrome smoke on :5182
mounts cleanly with no console errors.

Gate-Record: .workflow/records/1422-god-datapreview-bottompanel.json
Task-Kind: refactor
Issue: #1422
