+++ .workflow/records/1422-god-datapreview-bottompanel.json
@@ -1,0 +1,220 @@
+{
+  "admin_labels": [],
+  "amendments": [
+    {
+      "approved_by": null,
+      "exclude": [],
+      "include": [
+        "frontend/src/components/DataPreview.test.tsx",
+        "frontend/src/components/DataPreview.parts/refEntries.test.ts",
+        "frontend/src/components/BottomPanel.parts/codeBlockPorts.test.ts"
+      ],
+      "reason": "Add DataPreview.test.tsx to scope (touch imports) + test files for new parts modules"
+    }
+  ],
+  "branch": "refactor/issue-1422/god-datapreview-bottompanel",
+  "changed_test_paths": [
+    "frontend/src/components/DataPreview.parts/refEntries.test.ts",
+    "frontend/src/components/BottomPanel.parts/codeBlockPorts.test.ts"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "npm run lint (from frontend/)",
+      "exit_code": 0,
+      "name": "frontend-lint",
+      "output_path": "0 errors, 46 warnings (down from 47 pre-PR; all warnings pre-existing across other files except no-array-index-key/react-refresh which mirror originals)",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "npm run format:check (from frontend/)",
+      "exit_code": 0,
+      "name": "frontend-format",
+      "output_path": "All matched files use Prettier code style",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "npm run typecheck (from frontend/)",
+      "exit_code": 0,
+      "name": "frontend-typecheck",
+      "output_path": "tsc --noEmit clean (no output)",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "npx vitest run (from frontend/)",
+      "exit_code": 0,
+      "name": "frontend-test",
+      "output_path": "Test Files 46 passed (46); Tests 504 passed | 13 skipped (517); new files: BottomPanel.parts/codeBlockPorts.test.ts (15 cases), DataPreview.parts/refEntries.test.ts (8 cases)",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "npm run build (from frontend/)",
+      "exit_code": 0,
+      "name": "frontend-build",
+      "output_path": "built in 17.54s; 2101 modules transformed",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "vite dev :5182 + chrome-mcp navigation",
+      "exit_code": 0,
+      "name": "chrome-smoke",
+      "output_path": "App mounts cleanly at http://localhost:5182/, React root has content, no console errors. Local screenshot: docs/audit/smoke-test-1422-w2b-20260522.png (not committed)",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "PYTHONPATH=src python -m scistudio.qa.audit.full_audit --repo-root . --format json --output docs/audit/full-audit-latest.json",
+      "exit_code": 0,
+      "name": "full_audit",
+      "output_path": "docs/audit/full-audit-latest.json",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "addendum": {
+      "not_applicable": true,
+      "rationale": "N/A"
+    },
+    "adr": {
+      "not_applicable": true,
+      "rationale": "Behavior-preserving refactor; no ADR/spec change"
+    },
+    "changelog": {
+      "paths": [
+        "CHANGELOG.md"
+      ]
+    },
+    "checklist": {
+      "not_applicable": true,
+      "rationale": "Manager (frontend-cleanup-mgr) owns docs/planning/frontend-cleanup-1412-checklist.md; W2-B row updated by manager post-merge per checklist convention"
+    },
+    "docs": {},
+    "spec": {
+      "not_applicable": true,
+      "rationale": "N/A \u2014 no public contract change"
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "PYTHONPATH=src python -m scistudio.qa.audit.full_audit --repo-root . --format json --output docs/audit/full-audit-latest.json",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "docs/audit/full-audit-latest.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": false,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1422,
+      "url": "https://github.com/zjzcpj/SciStudio/issues/1422"
+    }
+  ],
+  "owner_directive": "Split DataPreview.tsx (1110 LOC, 38 hooks) and BottomPanel.tsx (1030 LOC) so every produced file is <=500 LOC. Behavior-preserving refactor; no API/UX changes.",
+  "planned_files": [
+    "frontend/src/components/DataPreview.tsx",
+    "frontend/src/components/DataPreview.parts/luts.ts",
+    "frontend/src/components/DataPreview.parts/ImageViewer.tsx",
+    "frontend/src/components/DataPreview.parts/TableViewer.tsx",
+    "frontend/src/components/DataPreview.parts/PreviewRenderer.tsx",
+    "frontend/src/components/DataPreview.parts/refEntries.ts",
+    "frontend/src/components/DataPreview.parts/useSlicePreview.ts",
+    "frontend/src/components/DataPreview.parts/useOmeMetadata.ts",
+    "frontend/src/components/BottomPanel.tsx",
+    "frontend/src/components/BottomPanel.test.tsx",
+    "frontend/src/components/BottomPanel.parts/CaretPreservingTextInput.tsx",
+    "frontend/src/components/BottomPanel.parts/FormatCapabilityConfig.tsx",
+    "frontend/src/components/BottomPanel.parts/codeBlockPorts.ts",
+    "frontend/src/components/BottomPanel.parts/CodeBlockConfigEditor.tsx",
+    "frontend/src/components/BottomPanel.parts/ConfigPanel.tsx",
+    "frontend/src/components/BottomPanel.parts/LogViewer.tsx",
+    "frontend/src/components/BottomPanel.parts/TabBar.tsx"
+  ],
+  "pull_request": null,
+  "record_path": ".workflow/records/1422-god-datapreview-bottompanel.json",
+  "required_checks": [
+    "frontend-lint,frontend-format,frontend-typecheck,frontend-test,frontend-build,sentrux"
+  ],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "frontend/src/components/DataPreview.tsx",
+      "frontend/src/components/DataPreview.parts/**",
+      "frontend/src/components/BottomPanel.tsx",
+      "frontend/src/components/BottomPanel.test.tsx",
+      "frontend/src/components/BottomPanel.parts/**",
+      "CHANGELOG.md",
+      ".workflow/records/1422-god-datapreview-bottompanel.json"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "sentrux check",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": "rules_checked=3/15 (all pass), quality_signal=4442, files=1164, lines=299223, import_edges=2650, no violations introduced (baseline quality_signal=4442 from manager bootstrap unchanged)",
+    "pro_required": false,
+    "quality_signal": null,
+    "rules_checked": null,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {},
+    "total_rules_defined": null
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "implement",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "update_docs",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "pending",
+      "summary": null
+    }
+  ],
+  "task_id": "1422-god-datapreview-bottompanel",
+  "task_kind": "refactor"
+}
