+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- [#1422] Split the two largest god files in the bottom panel surface — `frontend/src/components/DataPreview.tsx` (1110 LOC, 38 hooks) and `frontend/src/components/BottomPanel.tsx` (1030 LOC) — into smaller behavior-preserving modules under sibling `DataPreview.parts/` and `BottomPanel.parts/` directories. New `DataPreview.parts/`: `refEntries.ts` (`extractRefEntries`, `RefEntry`, `deriveDisplayName`), `luts.ts` (LUT colormap library + `applyLUTToImage`), `ImageViewer.tsx` (pan/zoom + LUT controls; further split into `useImagePanZoom`, `useLUTProcessing`, `ImageCanvas`, `SliceSliderRow`, `ZoomRow`, `LUTSelector`, `MinMaxRange` sub-components), `TableViewer.tsx` (`useTablePageFetch`, `readTablePayload`, `TableHeader`, `TableBody`, `PaginationControls`), `PreviewRenderer.tsx` (`switch` dispatcher split into per-`kind` render helpers), `useSlicePreview.ts` (#899 slice debounce + slice cache + stale-fallback hook), `useOmeMetadata.ts` (ADR-043 FR-013 OME panel hook). New `BottomPanel.parts/`: `CaretPreservingTextInput.tsx` (#710 caret-restore input), `FormatCapabilityConfig.tsx`, `codeBlockPorts.ts` (pure helpers + types: `CODEBLOCK_DATA_TYPES`, `codeBlockFolder`, `normalizeCodeBlockPort`, `persistCodeBlockPort`, `isCodeBlockConfigTarget`), `CodeBlockConfigEditor.tsx` (script/interpreter form + `CodeBlockEnvironmentEditor` + `CodeBlockPortTable`), `ConfigPanel.tsx` (variadic ports + format capability + dynamic schema fields, with `EnumField`/`ScalarField`/`ConfigField` extraction and `runBrowseDialog` helper), `LogViewer.tsx`, `TabBar.tsx` (`ALL_TABS`, `TAB_LABELS`, pin toggle). The slim `DataPreview.tsx` (126 LOC) and `BottomPanel.tsx` (108 LOC) now own only orchestration + JSX composition; their public exports are unchanged (`DataPreview`, `extractRefEntries`, `RefEntry`, `BottomPanel`). Behavior is exactly preserved: same canvas integration, same WebSocket events, same `data-testid` markers (`image-info-badge`, `image-slice-slider-row`, `image-slice-slider`, `unread-badge-*`, `bottom-panel-pin-toggle`, `open-ome-metadata`), same #899 slice debounce + stale-fallback semantics, same #710 caret-preservation guard, same ADR-038 §3.8 / ADR-039 §3.5 / ADR-043 FR-013 wiring. Three `eqeqeq` violations that were previously tolerated by the `BottomPanel.tsx` / `DataPreview.tsx` waiver entries are rewritten to strict equality (`x == null` → `x === null || x === undefined`); the original waiver block in `frontend/eslint.config.js` is untouched on this PR so the slim parent files stay covered (Wave 4 will strip the entries once all three Wave 2 PRs land). Tests: existing `BottomPanel.test.tsx` (14 cases) and `DataPreview.test.tsx` (8 cases) pass unchanged; new `BottomPanel.parts/codeBlockPorts.test.ts` (15 cases) and `DataPreview.parts/refEntries.test.ts` (8 cases) cover the extracted pure-helper modules. `npm run lint` reports 0 errors and 46 unchanged pre-existing warnings (down from 47 by removing dead `eslint-disable-next-line no-console` directives that the strict `no-console: { allow: ["warn","error"] }` rule no longer needs). `npm run format:check`, `npm run typecheck`, `npm test` (504 pass, 13 skipped), `npm run build` all clean. Chrome smoke test on port 5182 mounts the app without console errors. Parent agents W2-A (`App.tsx` + `BlockNode.tsx`) and W2-C (`RunDetail.tsx` + `lib/api.ts` + `ConflictMarkerDecoration.ts`) split the other god files in parallel against `track/frontend-cleanup-1412`. (@claude, 2026-05-22, branch: refactor/issue-1422/god-datapreview-bottompanel)
+
 - [#1408] Downgrade Sentrux from a hard gate to an advisory signal (ADR-042 Addendum 3). The Sentrux binary is not installable in many contributor and AI-agent execution environments (distributed as Rust source under the Claude marketplace; no prebuilt binary), so every source-/workflow-/architecture-/governance-touching PR was hitting `gate-record.sentrux.missing` at the local pre-push hook and at CI's workflow-gate, requiring an `admin-approved:ai-override` label purely for tooling reasons (worked example: PR #1406 shipped with the override label even though no `src/scistudio/` was touched). Two changes: (1) `src/scistudio/qa/governance/gate_record.py::validate_gate_record` no longer emits `gate-record.sentrux.missing` — missing, skipped, and unknown sentrux evidence are all allowed. The narrower `gate-record.sentrux.not-passing` finding is kept and tightened to fire only on recorded `status="fail"` (status="unknown" and status="skipped" no longer count) so the gate's remaining scope is the case it was actually designed for: a developer who ran sentrux, observed a real failure, and is pushing anyway. The override label and `--bypass-label` flag still work for justified exceptions. (2) `.github/workflows/workflow-gate.yml` calls `sentrux_gate.verify_free_tier_claims` unchanged but converts findings to `::warning::` annotations instead of adding them to the blocking reports list — the workflow-gate job never fails on sentrux alone, and reviewers read the advisories in PR check details. The supersession is narrow: only Addendum 1 §D1 (sentrux required as gate evidence) and §D5 (free-tier-honesty rule's blocking semantics) are affected. §D2 committed gate records, §D3 six-stage gate, §D7 implementation tests required, §D8 legacy gate removal, §D9 generated artifacts, §D10 AI gate CLI requirement, and §D11/§D12 architecture frontmatter / drift checks are unaffected. The Sentrux tool, `SentruxEvidence` data model, `gate_record sentrux` CLI subcommand, `verify_free_tier_claims`, and `.sentrux/rules.toml` all continue to exist; only the enforcement is downgraded. Tests reflect the inverted semantics: `tests/qa/test_gate_record_ci.py::test_sentrux_required_for_applicable_changes` renamed to `test_sentrux_missing_evidence_is_allowed_for_applicable_changes` (asserts missing does NOT block); two new tests `test_sentrux_skipped_evidence_is_allowed_for_applicable_changes` and `test_sentrux_recorded_failure_still_blocks_locally` cover the new threshold. Existing `test_sentrux_evidence_rejects_*` tests in `tests/qa/test_gate_record.py` retained unchanged because the `SentruxEvidence` validation model itself is unchanged (Pro-required claims and impossible rule counts still rejected at parse time). (@agent, 2026-05-21, branch: feat/sentrux-advisory)
 ### Fixed
