+++ frontend/src/components/BottomPanel.parts/LogViewer.tsx
@@ -1,0 +1,43 @@
+import { useMemo, useState } from "react";
+
+import type { LogEntry } from "../../types/api";
+
+export function LogViewer({ entries }: { entries: LogEntry[] }) {
+  const [level, setLevel] = useState("all");
+  const filtered = useMemo(() => {
+    return entries.filter((entry) => level === "all" || entry.level === level);
+  }, [entries, level]);
+
+  return (
+    <div className="flex h-full flex-col">
+      <div className="mb-4 flex items-center gap-3">
+        <select
+          className="rounded-full border border-stone-300 bg-white px-3 py-2 text-sm"
+          onChange={(event) => setLevel(event.target.value)}
+          value={level}
+        >
+          <option value="all">All levels</option>
+          <option value="info">Info</option>
+          <option value="error">Error</option>
+        </select>
+      </div>
+      <div className="flex-1 overflow-auto rounded-[1.4rem] border border-stone-200 bg-stone-950 p-4">
+        {filtered.length ? (
+          filtered.map((entry, index) => (
+            <div
+              className="border-b border-stone-800 py-2 text-sm text-stone-100"
+              key={`${entry.timestamp}-${index}`}
+            >
+              <p className="text-[11px] uppercase tracking-[0.3em] text-stone-500">
+                {entry.level} · {entry.workflow_id ?? "workflow"} · {entry.block_id ?? "system"}
+              </p>
+              <p className="mt-1">{entry.message}</p>
+            </div>
+          ))
+        ) : (
+          <p className="text-sm text-stone-500">No logs yet.</p>
+        )}
+      </div>
+    </div>
+  );
+}
